requirements.txt
openai
pandas
pycaret
python-dotenv
streamlit
